LPS-102145 Reduce duplication.
