+++ modules/apps/change-tracking/change-tracking-web/src/main/java/com/liferay/change/tracking/web/internal/portlet/ChangeListsPortlet.java
@@ -19,17 +19,9 @@ import com.liferay.change.tracking.service.CTCollectionLocalService;
 import com.liferay.change.tracking.service.CTEntryLocalService;
 import com.liferay.change.tracking.service.CTPreferencesLocalService;
 import com.liferay.change.tracking.service.CTProcessLocalService;
-import com.liferay.change.tracking.web.internal.configuration.CTConfiguration;
 import com.liferay.change.tracking.web.internal.constants.CTWebKeys;
 import com.liferay.change.tracking.web.internal.display.context.ChangeListsDisplayContext;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-import com.liferay.portal.kernel.model.Role;
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.kernel.security.auth.PrincipalException;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.security.permission.ResourceActions;
-import com.liferay.portal.kernel.security.permission.UserBag;
 import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.servlet.SessionMessages;
@@ -38,18 +30,11 @@ import com.liferay.portal.kernel.util.Portal;
 
 import java.io.IOException;
 
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-
 import javax.portlet.Portlet;
 import javax.portlet.PortletException;
-import javax.portlet.PortletRequest;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 import org.osgi.service.component.annotations.Reference;
@@ -81,7 +66,7 @@ import org.osgi.service.component.annotations.Reference;
 	},
 	service = Portlet.class
 )
-public class ChangeListsPortlet extends MVCPortlet {
+public class ChangeListsPortlet extends BaseChangeListPortlet {
 
 	@Override
 	public void render(
@@ -115,46 +100,6 @@ public class ChangeListsPortlet extends MVCPortlet {
 
 		super.render(renderRequest, renderResponse);
 	}
-
-	@Activate
-	protected void activate(Map<String, Object> properties) {
-		Set<String> administratorRoleNames = new HashSet<>();
-
-		CTConfiguration ctConfiguration = ConfigurableUtil.createConfigurable(
-			CTConfiguration.class, properties);
-
-		Collections.addAll(
-			administratorRoleNames, ctConfiguration.administratorRoleNames());
-
-		_administratorRoleNames = administratorRoleNames;
-	}
-
-	@Override
-	protected void checkPermissions(PortletRequest portletRequest)
-		throws Exception {
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		if (permissionChecker.isCompanyAdmin()) {
-			return;
-		}
-
-		UserBag userBag = permissionChecker.getUserBag();
-
-		for (Role role : userBag.getRoles()) {
-			if (_administratorRoleNames.contains(role.getName())) {
-				return;
-			}
-		}
-
-		throw new PrincipalException(
-			String.format(
-				"User %s must have administrator role to access %s",
-				permissionChecker.getUserId(), getClass().getSimpleName()));
-	}
-
-	private Set<String> _administratorRoleNames;
 
 	@Reference
 	private ClassNameLocalService _classNameLocalService;
